Onclicks added, design updated
